+++ np.CallbackURLs/CHANGELOG.md
@@ -4,13 +4,19 @@
 
 See Plugin [README](https://github.com/NotePlan/plugins/blob/main/np.CallbackURLs/README.md) for details on available commands and use cases.
 
+## [1.3.0] - 2022-12-21 (@dwertheimer)
+
+- Fix bug when selecting self-running template
+- Include ability to create self-running template
+- Hide x-success behind a preference field
+
 ## [1.2.1] - 2022-12-08 (@dwertheimer)
 
 - @jgclark changed self-running templates to use semicolons to separate variables. Updated the URL maker to match
 
 ## [1.2.0] - 2022-12-04 (@dwertheimer)
 
-- Added links to lines 
+- Added links to lines
 - Updated the way hashtags in titles are encoded (which changed in NotePlan) -- strip out hashtags in headings
 
 ## [1.1.2] - 2022-10-02 (@dwertheimer)